day 11 - caught up!
Day 11 was cool, navigating seats and validating rules.

Initially my solution was iterating the whole array for each potential neighboring seat, but for part 2 I realized that'd be massively inefficient. Implemented a Dictionary that mirrored the structure of the input file, but instead of a token, the spot class was in place. Made finding and navigating the array very easy :)